+++ docs/CHANGELOG.md
@@ -3,6 +3,262 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.3.45 (2020-04-23)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.44 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.43 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.42 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.41 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.40 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.39 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.38 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.37 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.36 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.35 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.34 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.33 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.32 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.31 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.30 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.29 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.28 (2020-04-22)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.27 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.26 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.25 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.24 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.23 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.22 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.21 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.20 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.19 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.18 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.17 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.16 (2020-04-21)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.15 (2020-04-20)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
+## 1.3.14 (2020-03-24)
+
+**Note:** Version bump only for package @accera/solar-components.docs
+
+
+
+
+
 ## 1.3.13 (2020-03-16)
 
 **Note:** Version bump only for package @accera/solar-components.docs